Echo-bot/requirements.txt
python-telegram-bot==21.10
python-dotenv==1.0.1